2003-02-03  Patrick Eriksson  <Patrick.Eriksson@rss.chalmers.se>
	* arts-1-1-439

	* src/ppath.cc: I found more new numerical problems with Numeric=float.
	This time for calculations with refrcation. It turned out that step
	lengths of e.g. 1 km could not be used. Decided then to test to
	hard-code double as far as possible in the function. Everything
	compiles without warnings. Not yet checked if things work better for
	Numeric=float.

	* src/math_funcs.h: Removed declaration of non-existing return version
	of function of type linspace.

	* src/m_basic_types.cc (VectorNLogSpace): Not using removed return
	version of nlogspace.

	* src/math_funcs.cc (nlinspace): The last value is now set to be stop
	exactly (not start + i*strep).

	Removed old return version of nlogspace.


git-svn-id: https://www.sat.ltu.se/svn/rt/arts/trunk@2719 aaf1aab0-4228-0410-ad68-8dceda47f409

+++ src/m_ppath.cc
@@ -199,7 +199,7 @@ void DoGridcell2D(
   const Numeric   ppc = geometrical_ppc( r_start, za_start );
 
   Vector    r, lat, za;
-  Numeric   lstep;
+  double    lstep;
   Index     endface, tanpoint;
 
   do_gridcell_2d( r, lat, za, lstep, endface, tanpoint, r_start, lat_start, 
@@ -219,7 +219,7 @@ void DoGridcell2D(
   xml_write_to_file ( filename, za );
   filename = "";
   filename_xml( filename, "lstep" );
-  xml_write_to_file ( filename, lstep );
+  xml_write_to_file ( filename, Numeric(lstep) );
   filename = "";
   filename_xml( filename, "endface" );
   xml_write_to_file ( filename, endface );
@@ -261,7 +261,7 @@ void DoGridcell3D(
   const Numeric   ppc = geometrical_ppc( r_start, za_start );
 
   Vector    r, lat, lon, za, aa;
-  Numeric   lstep;
+  double    lstep;
   Index     endface, tanpoint;
 
   do_gridcell_3d( r, lat, lon, za, aa, lstep, endface, tanpoint, 
@@ -289,7 +289,7 @@ void DoGridcell3D(
   xml_write_to_file ( filename, aa );
   filename = "";
   filename_xml( filename, "lstep" );
-  xml_write_to_file ( filename, lstep );
+  xml_write_to_file ( filename, Numeric(lstep) );
   filename = "";
   filename_xml( filename, "endface" );
   xml_write_to_file ( filename, endface );
